+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v1.0.2
 #### Updated
+- Updating Results.jsx to remove the unnecessary function to parse the snippet of a result item.
 - Tab attributes use tab names instead of numbers
 - The current tab is tracked from the selectedFacet property in the Store
 #### Added