+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Changelog
 
-### v1.0.2
+### v1.1.0
 #### Updated
 - Updadting Results.jsx to move the functions rerlated to modeling data out of it.
 - Updating Results.jsx to remove the unnecessary function to parse the snippet of a result item.
@@ -8,6 +8,7 @@
 - The current tab is tracked from the selectedFacet property in the Store
 #### Added
 - Adding the instruction message to ask the user to enter a search keyword.
+- Adding the methods to call AWS KMS service to decrypt the API root for search requests.
 
 ### v1.0.1
 #### Updated